refactor: split docs and node wrapper
