+++ scripts/vite/paths.ts
@@ -1,0 +1,37 @@
+import { dirname, resolve } from "node:path";
+import { fileURLToPath } from "node:url";
+
+/** Absolute repository root derived from this config helper's file URL. */
+export const rootDir = resolve(dirname(fileURLToPath(import.meta.url)), "../..");
+
+/** Native napi package directory used by wrapper build tasks and aliases. */
+export const nodePackageDir = resolve(rootDir, "src/bindings/nodejs/corsa_node");
+
+/** Type-aware Oxlint package directory used by pack and alias configuration. */
+export const corsaOxlintDir = resolve(rootDir, "src/bindings/nodejs/corsa_oxlint");
+
+/** Source root for the `corsa-oxlint` package pack task. */
+export const corsaOxlintSourceDir = resolve(corsaOxlintDir, "ts");
+
+/** Files generated by the native Node wrapper build and ignored by local tooling. */
+export const generatedNodeArtifacts = [
+  "src/bindings/nodejs/corsa_node/index.d.ts",
+  "src/bindings/nodejs/corsa_node/index.js",
+  "src/bindings/nodejs/corsa_node/ts/**/*.d.ts",
+  "src/bindings/nodejs/corsa_node/ts/**/*.js",
+  "src/bindings/nodejs/corsa_node/ts/**/*.js.map",
+];
+
+/** Upstream checkouts are inputs, not code owned by this workspace. */
+export const upstreamRefPatterns = ["origin/**", "ref/**"];
+
+/** Ignore list shared by JS lint and format tasks. */
+export const lintIgnorePatterns = [
+  ...generatedNodeArtifacts,
+  ...upstreamRefPatterns,
+  "bench/fixtures/**",
+  "src/bindings/nodejs/corsa_node/ts/**/*.test.ts",
+];
+
+/** No-op command used by aggregate Vite+ tasks. */
+export const noopCommand = 'node -e "process.exit(0)"';
