+++ src/tools/corsa_docs/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "corsa_docs"
+version = "0.1.0"
+edition = "2021"
+license.workspace = true
+publish = false
+repository.workspace = true
+rust-version.workspace = true
+
+[dependencies]
+ox_content_allocator = "2.14.0"
+ox_content_docs = "2.14.0"
+ox_content_parser = "2.14.0"
+ox_content_renderer = "2.14.0"
+ox_content_ssg = "2.16.0"
+thiserror = "2"
